backend/requirements.txt
absl-py==2.2.2
astunparse==1.6.3
blinker==1.9.0
cachetools==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
Flask==2.3.2
Flask-Cors==4.0.0
flatbuffers==25.2.10
gast==0.6.0
google-auth==2.40.2
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.71.0
gunicorn==23.0.0
h5py==3.13.0
idna==3.10
importlib-metadata==8.7.0
itsdangerous==2.2.0
jinja2==3.1.6
keras==2.14.0
libclang==18.1.1
markdown==3.8
MarkupSafe==3.0.2
ml-dtypes==0.2.0
numpy==1.26.4
oauthlib==3.2.2
opt-einsum==3.4.0
packaging==25.0
pillow==11.2.1
protobuf==4.25.7
pyasn1==0.6.1
pyasn1-modules==0.4.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-intel==2.14.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.1.0
typing-extensions==4.13.2
urllib3==2.4.0
werkzeug==3.1.3
wrapt==1.14.1
zipp==3.21.0
